+++ src/main/java/me/lokka30/microlib/ItemBuilder.java
@@ -31,7 +31,9 @@ public class ItemBuilder {
     private int amount = 1;
     private int damage = 0;
     private Map<Enchantment, Integer> enchantments = new HashMap<>();
+    @SuppressWarnings("CanBeFinal")
     private List<String> lore = new ArrayList<>();
+    @SuppressWarnings("CanBeFinal")
     private List<ItemFlag> itemFlags = new ArrayList<>();
 
     public ItemBuilder(Material material) {