+++ src/main/java/me/lokka30/microlib/UpdateChecker.java
@@ -20,8 +20,8 @@ import java.util.Scanner;
 @SuppressWarnings("unused")
 public class UpdateChecker {
 
-    private JavaPlugin plugin;
-    private int resourceId;
+    private final JavaPlugin plugin;
+    private final int resourceId;
 
     public UpdateChecker(JavaPlugin plugin, int resourceId) {
         this.plugin = plugin;