+++ src/main/java/me/lokka30/microlib/events/PlayerMoveFullXYZEvent.java
@@ -16,8 +16,8 @@ import org.bukkit.event.HandlerList;
 @SuppressWarnings("unused")
 public class PlayerMoveFullXYZEvent extends Event {
     private static final HandlerList HANDLERS = new HandlerList();
+    private final Player player;
     private boolean isCancelled = false;
-    private Player player;
     private Location from, to;
 
     public PlayerMoveFullXYZEvent(Player player, Location from, Location to) {
@@ -52,5 +52,13 @@ public class PlayerMoveFullXYZEvent extends Event {
 
     public Location getTo() {
         return to;
+    }
+
+    public void setFrom(Location from) {
+        this.from = from;
+    }
+
+    public void setTo(Location to) {
+        this.to = to;
     }
 }